feat: animation, slogan
